Moving all loader configs to a separate file. run.js is going out of hand otherwise
